Split service into separate package
