+++ mlflow/go/pkg/service/model.go
@@ -1,0 +1,5 @@
+package service
+
+import "github.com/mlflow/mlflow/mlflow/go/pkg/contract"
+
+var ModelRegistryService contract.ModelRegistryService
